+++ include/GafferScene/ShaderTweaks.h
@@ -74,6 +74,9 @@ class GAFFERSCENE_API ShaderTweaks : public AttributeProcessor
 		void hashProcessedAttributes( const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
 		IECore::ConstCompoundObjectPtr computeProcessedAttributes( const Gaffer::Context *context, const IECore::CompoundObject *inputAttributes ) const override;
 
+		// Overridden from AttributeProcessor so that we can tweak shaders stored as options as well as attributes.
+		IECore::ConstCompoundObjectPtr computeGlobals( const Gaffer::Context *context, const ScenePlug *parent ) const override;
+
 		static size_t g_firstPlugIndex;
 
 	private :