+++ python/GafferSceneTest/ShaderTweaksTest.py
@@ -131,6 +131,32 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		tweaks["localise"].setValue( True ) # Nothing to localise, so should be same as before
 		self.assertEqual( tweaks["out"].globals()["attribute:surface"].outputShader().parameters["i"], IECore.IntData( 12 ) )
 
+	def testOption( self ) :
+
+		shader = GafferSceneTest.TestShader()
+		shader.loadShader( "simpleShader" )
+
+		renderPassShader = GafferScene.RenderPassShader()
+		renderPassShader["shader"].setInput( shader["out"] )
+		renderPassShader["usage"].setValue( "shadowCatcher" )
+		renderPassShader["renderer"].setValue( "Cycles" )
+
+		self.assertIn( "option:renderPass:shader:shadowCatcher:Cycles", renderPassShader["out"].globals() )
+		self.assertEqual( renderPassShader["out"].globals()["option:renderPass:shader:shadowCatcher:Cycles"].outputShader().parameters["i"], IECore.IntData( 0 ) )
+
+		tweaks = GafferScene.ShaderTweaks()
+		tweaks["in"].setInput( renderPassShader["out"] )
+		tweaks["global"].setValue( True )
+		tweaks["shader"].setValue( "renderPass:shader:shadowCatcher:Cycles" )
+
+		self.assertEqual( tweaks["out"].globalsHash(), tweaks["in"].globalsHash() )
+		self.assertEqual( tweaks["out"].globals(), tweaks["in"].globals() )
+
+		tweaks["tweaks"].addChild( Gaffer.TweakPlug( "i", 2 ) )
+		tweaks["tweaks"][0]["mode"].setValue( Gaffer.TweakPlug.Mode.Add )
+		self.assertNotEqual( tweaks["out"].globalsHash(), tweaks["in"].globalsHash() )
+		self.assertEqual( tweaks["out"].globals()["option:renderPass:shader:shadowCatcher:Cycles"].outputShader().parameters["i"], IECore.IntData( 2 ) )
+
 	def testSerialisation( self ) :
 
 		s = Gaffer.ScriptNode()